bug fixes during coverage audit
Mining the code coverage report for ideas on new test cases led to
discovering 3 unreported bugs.

1. When performing a conditional replace on an unbounded cache, the new
value was supplied to the removal listener. The intent is for the old
value to be sent.
2. When computing an entry that replaces one that is evictable but not
yet removed from the underlying map, the associated removal notification
was not always published.
3. During Map#equals the traversal may remove entries pending eviction.
If the subset of live entries are equivalent then it could incorrectly
return true. Instead when an eviction is triggered then the size
prescreen is no longer accurate and should return false.