Avoid overwrites due to codegen compilation
The per-entry classes are code generated for optimizing the field layout
(memory overhead) and compiled without debug symbols to minimize the jar
size. Previously this was a double compilation due to circular
dependencies. That busted the build cache and made it harder to detect
if from a mallicious supply chain attack.

The factory classes are no longer generated as simple logic. Previously
they were due to needing to build a very large switch to each class, but
has long since been switched to a reflection lookup. That reduces the
factory class load time and jar bloat. This means that the hand written
classes have no static dependencies on the generated classes, and the
two can be independent compilation units.

The compileCodeGenJava task was run twice due to Gradle rewriting
errorprone's dependency as a local jar dependency. This was only
partially worked around where the main project downloads a local jar
to trick the classpath, but other subprojects included the build's jar.
This is now excluded on the dependency so the cache shouldn't be
busted.

As part of the build cleanup, the pmd checks for JMH were enabled and
the false warnings suppressed.