reduce build time when publishing to Sonatype
Gradle does not cache jar or distribution tasks, which is usually
fast but slow for the simulator due to many dependencies. The shadow
plugin increases this build time as it generates dependency free
variants. This case can be removed as it comes in via jmh, for an
experiment to validate a back-of-the-envelope estimate. As that is
no longer needed we can treat it as dead code and shave off 30s.