Display test results as a badge and action job summary
A single build executes ~9 million tests in order to cover the
api, feature matrix, linearization, Guava compatibility, and 98^
code coverage. A build is performed on JDK 11 and 18 (x86) and
17 (arm64), where the ARM run assert that there are no latent
bugs due to a weaker memory model. This takes ~2.5 hours despite
using parallel tests and jobs.

As it is hard to track our test size, the JDK 18 run is used as
the reference. The test results are aggregated and published to
have some visibility. We do not count the JDK 11 x86 and JDK 17
arm64 runs as repeats.