Add Coherence to analysis suite
The next release of Coherence (expected this month) will adopt Caffeine
and advocate it as the preferred local cache. The current in-memory
cache is retained for backwards compatibility (a 20yr old code base).

Coherence is one of the most well respected distributed data grids and
used heavily at large enterprises (banks, train logistics, etc). While
their cache mitigates many bottlenecks due to being distributed and
partitioned, they still observe problems at scale. An internal customer
at Oracle is making the switch due to the high rate of evictions and
coarse write locking causing their application to block and timeout on
puts.

An analysis of the eviction throughput shows that Caffeine is 2.9x to
2050x faster, depending on the cache size (1 - 10M entries). Caffeine
evicts in the low millions of entries per second, which is more than
enough for real workloads.

An analysis of the concurrent read/write throughput shows Caffeine
being 3.8x faster on reads 17.8x faster on writes. This was performed
on an old 4 core macbook for a rough estimate.

The hit rate improvement ranges from 8% to 33% on a variety of real
workloads. A higher hit rate results in fewer writes, fewer evictions,
and less garbage generated (reducing GC overhead).

This data indicates that we should be able to solve or greatly alleviate
their customer's performance problems.