Strengthen removeAll, removeIf, and retainAll in Map views
The inherited implementation from AbstractCollection uses the
iterator for removing on a match. The iterator remove does so
by key and is unconditional. In a concurrent collection the
underlying entry may be modified so that it no longer matches,
making the removal unexpected.

ConcurrentHashMap uses this iterator approach for removeAll and
retainAll, and is conditional for removeIf. Most others like Guava
likewise inherit the iterator approach. For now Caffeine will be
stronger in its evaluations, but users should consider this to be
undefined behavior and not rely on it given that others are less
strict.