Fix gradle npe when reporting test results
When a test runs with expected exceptions that are never thrown then
TestNG provides the listener with the test results as if having run
successfully but with an TestException. After the listener the test
results are updated to indicate a test failure and this is reported
to Gradle's runner. Due to our memory optimizations which think that
the test passed, the exception was discarded. This caused Gradle to
fail with an NPE as it is not defensive against a failue without a
cause (gradle/gradle#21336).

Since we run million of tests the results being held in-memory is not
feasible. The exceptions are lightweight though, so we don't need to
discard them to avoid out-of-memory errors.

Also fixed sonarqube and other minor touchups while resolving this
problem.