Improve eviction when overflow or the weight is oversized (#745)
Previously an attempt to centralize evictions into one code path led to
a suboptimal approach (464bc19). This tried to move those entries into
the LRU position for early eviction, but was confusing and could too
aggressively evict something that is desirable to keep. We now follow
the following logic for exceptional cases,

- If the entry exceeds the cache's maximum weight, then it is evicted
  immediately.
- If the entry is in the window and exceeds the window's maximum weight,
  then move it to the LRU positiion for immediate promotion. Otherwise
  the window will be flushed anyways to eject this entry, so a promotion
  decision can be done earlier with less impact on other window entries.
- If the cache's weight exceeds the implementation's overflow limit
  (2^63 - 2^31), then trigger an eviction before processing other
  events. This avoids a weight (long) overflowing and becoming negative.

The hit rate impact for weighted traces ranged from -2% to +4%, so the
changes do not cause a significant regression and are within a noise
tolerance. These changes do not impact non-weighted workloads.