clarify the code's intent in the ring buffers
These changes have no visible performance impact but better align to the
expected behavior, rather than being implicitly correct.

The striped buffer would return a thread contention failed status if
initializing a new ring buffer, which is preallocated with the element.
The caller only checks if appending failed due to being full, so this
change to the correct result has no effect.

The ring buffer's write counter CAS is atomic and does not require any
acquire or release memory barrier ordering upon completion. A release
fence is performed when setting the array slot, so we can piggyback on
that. While this does not impact x86 (TSO), it could have a slight
difference on a weaker memory model (e.g. ARM).