Remove akka because it is no longer an open-source project
We restrict the promotion of community and project adoption to open-source
projects so as to not imply a commercial bias. Unfortunitly the recent
announcement by the akka project requires that we remove it from this
list and our own usage in the cache simulator.

The cache simulator used Akka for convenience as it models the
eviction policy as an actor. This allows the actors to be executed in
parallel, where each actor applies the access events in sequence. An
actor runs as a virtual thread by coordinating its work using
ForkJoinPool to avoid platform thread context switches.

We emulate this behavior by chaining a sequence of CompletableFuture
tasks per actor. The buffering of messages in the mailbox is managed
by the chain length, where a semaphore blocks the sending of new work
until the actor has caught up. This has the same performance, is less
code, and does not require learning a foreign framework.