asynchronously complete the simulations
A small oversight when replacing akka was that the finishing step was
sequential. This meant that the main driver enqueued the command and
waited for the policy complete, and then moved onto the next one. For
most policies this step only performs validation checks and is fast.
However, the optimal policy does all of its work in this final step as
it has to know the complete future in order to make its eviction
decisions. Now the shutdown commands are sent in parallel and the
driver waits for all to complete, which resolves this slight
performance regression in our analysis tool. While the akka version
is still slightly faster, likely due to optimizations around locks and
ForkJoinPool task submissions, the difference is mostly noise.