Optimize ehcache's usage in the simulator
This change results in the same hit rate and has a 37% faster runtime.
For the DS1 trace this reduces from 91 minutes to 67 minutes, which is
still abysmal compared 15-30 seconds of all other caches. Unfortunately
this issue has been known and rejected since 2015. A simple change is
to hash the key to improve the distribution, resulting in a 1.4 minute
runtime, but it reduces the hit rate in for their sampled lru policy.