Optimize the CountMin4 sketch to fit in the cache line
Previously the frequency sketch distributed the counters uniformly across
the table. This approach tries to minimize the chance of collisions, but
means that each index is unpredictable and requires an additional memory
access to retreive. By constraining an item to a 64 byte cache line, all
of the counters can be selected with a single memory fetch. This change
substantially improves the throughput by up to 2.6x & 2.4x for the
frequency and increment methods. That can have a small benefit for cache
operations in a serial benchmark, e.g. a 17% improvement for evictions.
As the maintenance work is normally asynchronous, non-blocking, and
already very fast, the differences are unlikely to be observed in
practice.