adjust when explicit puts discard a refresh
Reviewing this code, I don't recall exactly why it was safe to discard the
refresh prior to the synchronized block. Likely it was because those other
conditions covered insert and putIfAbsent overwriting an expired entry,
whereas a live update will necessarily overwrite the value. Yet if a
putIfAbsent is overwriting an expired entry, that implies that it beat an
eviction which did not discard that in-flight future. As synchronization
might be delayed, it would still seem like an ABA style problem could
occur in a worst case scenario.

In all other updates the refresh is discarded within the synchronized
block. The refresh callback validate itself under this lock by removing
its future and, if successful, checking the results are consistent. It
seems safer to maintain this pattern on put, even if an unlikely or
perhaps impossible race, if only for consistency and avoid having to
rethink it all anew every few years when reviewing the code.