fix lincheck failing on jdk19 due to broken asm transformation
Lincheck's classloader will rewrite usages of ForkJoin's commonPool to an
instrumented instance. This requires opening additional modules for access,
or else a static initialization failure is thrown and the class is left in
an invalid state. This results in the test rerunning indefinitely. Once
this module restriction is lifted it is able to proceed as expected and
validate that the cache passes linearization tests.

https://github.com/Kotlin/kotlinx-lincheck/issues/130