defensively copy prior to JCache putAll (fixes #841)
Previuously, calling javax.cache.Cache.putAll(Map) with an immutable map
caused an UnsupportedOperationException if the CacheWriter throws an
exception. The problem with this is that you loose the original
exception from the CacheWriter and thus do not know what the root cause
was.

The underlying problem is that CacheWriter requires that the map
argument must be mutable so that it can communicate a partial success.
It appears that the TCK does not assert that mutation must be exposed
back to the Cache.putAll caller, so a defensive copy into a mutable map
satisfies the contracts and allows callers to use an immutable map.

Co-authored-by: Ken Dombeck <kdombeck@gmail.com>