ensure prompt clean up if a scheduler is configured (fixes #859)
A benign race can occur when a maintenance cycle is running, writers
insert new entries, and all of the entries that the policy is aware of
have expired so they are removed. In this case no clean up was scheduled
by the expiration policy, the maintenance status is left as "required",
and entries in the write buffer are waiting to be applied. If the cache
is idle then those entries will expire and the removal listener will not
be notified promptly. As a scheduler is configured to communicate that
a prompt notification is desired, the cache should more aggressively use
it to ensure this case is handled.