consistant behavior for null lookups to into unmodifiable maps (fixes #864)
While a null lookup is not allowed for cache the operations (e.g. asMap.get), if
an operation returned a map then this might not be consistant. Originally all of
these cases used Collections' emptyMap() or unmodifiableMap(m), but a few cases
switched to Map's of() or copyOf() alternatives. The former allows for null
queries whereas the latter does not, and this could vary depending on if the
result was populated. For consistency and retaining the original behavior, this
is restored to the Collections' methods. Further, the empty map is now wrapped
as unmodifiable due to the inconsistency otherwise with Map.clear()'s behavior.