Add @CheckReturnValue at the package level (fixes #863)
This annotation is used by ErrorProne and Spotbugs to warn users when
they ignore the results of a method call. This is now the default
behavior rather than a few opt in cases. In rare cases the result is
an ignorable signal, like List's add(e) or our LoadingCache.refresh(k),
so @CanIgnoreReturnValue is used to disable the check for those methods.
For all other methods then users can suppress the warning when desired.