modernize the gradle build
Previously the build used the Groovy DSL with script plugins. The Kotlin
DSL is now Gradle's default with a recommendation of using pre-compiled
script plugins. This provides an opportunity to clean up the build logic
by splitting our integrations into separate plugins. The dependencies
are now managed in version catalogs to allow for automatic updates.