Use jakarta.inject in jcache (fixes #1009)
The injection support is minimal to allow for JCache to load wired
classes, rather than use the default constructor that it would
otherwise assume.

The JSR provided annotation support for Spring, CDI, and Guice. These
only use javax.inject, so those users would need to support both
annotations, set our TypesafeConfigurator explicitly, or use byte code
rewriting (or a fork) with the new package. Guice provides dual support
in v6 and drops support in v7. As JCache is abandonware, it is likely
best to not use the annotations anyway.