Fix for #1065: Make EventDispatcher an instance field
This sidesteps an edge-case wherein a JCache listener in a primary cache
mutates a secondary Caffeine cache. While this is typically not a problem,
if the thread that mutates the primary cache is from a ForkJoinPool, the
EventDispatcher can deadlock as the calling thread may be re-used. Under
this condition, the secondary cache mutation is effectively waiting on its
own Future and will deadlock.