add tests for #1065 and make the jcache configuration fluent
The unit and integration tests verify that the event dispatcher does
not leak the pending acknowledgements across threads if the executor
descides to run the listener on the same thread as the publisher. This
occurred due to using a static ThreadLocal rather than per cache
instance, which was a mistake as typically that is how ThreadLocals
are meant to be used.

While crafting tests, it was noticed that the CaffeineConfiguration's
setters are not fluent like the JCache's basic implementation,
MutableConfiguration. This was changed for consistency and convenience.