added logging when an async cache's removal listener fails
The documented behavior when an exception is thrown is that the error
is logged and swallowed. However, the logging was only performed for a
synchronous cache. This inconsistency may have masked errors so it is
now fixed.

All other cases when logging is performed is now asserted, including
their message contents.