Fix containsKey for async cache's synchronous view when in-flight (fixes #1626)
The keySet and its iterator will now suppress in-flight entries for
the contains test and by its iterator. The retain and removal methods
will discard in-flight mappings, like the entrySet view does. This is
intentional as it is better to overly discard data then keep stale
contents due to incorrect linearization assumptions (of which async can
offer few). The synchronous views are best-effort, lenient, and try to
match a user's likely expected or intended behavior while being
conservative by being cautiously safe. Thus, small leaks of in-flight
behavior is preferrable, especially given the concurrent nature of the
use-case, and we encourage using the async asMap view when
orchastrating more nuanced behavior.