add a clairvoyant admission policy to the simulator
Like Bélády's optimal policy for eviction, this estimates the upper
bound for an admission policy by using future knowledge to make the
best decision. The candidate entry is only admitted if its next access
time is less than the victim's.

While the optimal replacement policy is able to dictate the working
set, the optimal admission policy is only able to advise whether to
keep the given candidate or victim entry. Thus, the eviction policy's
choices may result in a sequence where a more naive admission policy
outperforms the clairvoyant one. The intent is to show the upper
bounds of an adaptive admission policy based on its best guess rather
than the upper bound of the overall cache hit rate.