remove workaround for historic fork-join pool bugs fixed in jdk8
Issues like JDK 8u40 (JDK-8078490) and 8u152 (JDK-8200347) could cause
the dependent action to run twice, as observed in the debugger. It
seemed that the active task was stolen by another worker, which in this
case would only result in skewed statistics. These issues have long
since been resolved in late JDK-8 and this library required JDK-11 so
removable.