fix async cache's synchronous computeIfAbsent to retry if incomplete
AsyncCache provides a non-linearizable, atomic computeIfAbsent in the
synchronous view. This avoids long hold times of the Map's hashbin lock
by waiting on the pending future outside of the computation, inserting
its own if absent or the pior future holds a null value (thus pending
automatic removal). This retry loop attempt fails if the compute finds
an incomplete future, but previously it could replace that in a narrow
race condition due to a mistaken conditional. This case was caught when
adding test coverage for the putIfAbsent and computeIfAbsent retry
loops when this edge case occurs.