use parallel gc for parameterized tests
The junit xml reports from the slowest parameterized test tasks showed
the time spent per execution was 3 seconds in tests that did almost no
work. A JFR profile showed that G1 was performing thousands of GCs
and spent minutes in oldgen, with a max pause time of 3s. That task ran
in 16m38s for 1,370,023 tests.

Using ParallelGC reduced the time to 9m55s with 7s spent in GC and
very few old generation collections. This was already used on soft/weak
reference tests to have predictable reference collection.

The method profile shows the hotspots are within Truth assertions,
which is a known problem.